recipes-demos: ti-lvgl-demo: add high resolution OOB demo
LVGL is a library used for building GUIs for embedded devices. It is
light-weight, has a small footprint and does not necessarily require a
GPU to render its graphics. Since it works on Linux-capable processors,
it is a good choice for writing GUI applications for devices without a
GPU.

We have a new, high-resolution demo, written using LVGL. It is a part of
lvgl/demos directory. At present, it directly uses framebuffers to
display its graphics.

The features of this demo are:
    - "Smart Home", "EV Charging", "About" etc screens
    - Displaying live values from the ADC
    - Playing audio files
    - Controlling certain features, like audio volume, through phone,
      using MQTT

Integrate this demo into Yocto. This demo will be OOB experience for
certain devices, hence also write a systemd file that enables it to run
by-default.

Due to its audio and MQTT-connectivity features, it requires alsa-lib
and mosquitto packages. So add those as dependencies. To connect using
MQTT, it requires a certificate in `/usr/share/ti-lvgl-demo/`. So add
that as well.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>